refactor(settings): extract about and keybindings pages
Move the static reference pages out of the settings layer so it no longer
mixes catalog construction with about and shortcut content.
