+++ src/ui/settings.rs
@@ -18,11 +18,15 @@ use crate::{
     },
 };
 
+mod about;
 mod bindings;
 mod general;
+mod keybindings;
 mod theme;
+use about::about_page;
 use bindings::{bind_choice, bind_switch};
 use general::general_page;
+use keybindings::keybindings_page;
 use theme::theme_page;
 
 #[cfg(test)]
@@ -2100,151 +2104,6 @@ fn update_check_message(result: &UpdateCheck, update_method: UpdateMethod) -> St
             glib::markup_escape_text(message)
         ),
     }
-}
-
-fn keybindings_page(manager: Rc<ThemeManager>) -> gtk::Widget {
-    let content = page_content();
-    append_heading(&content, "KEYBINDING HINTS");
-    let (row, toggle) = settings_option(
-        "Show keybinding hints",
-        "Show navigation hints and paste availability at the bottom of every view. F1 opens the full reference even when hints are hidden.",
-        manager.show_keybinding_hints(),
-    );
-    bind_switch(
-        &manager,
-        &toggle,
-        ThemeManager::show_keybinding_hints,
-        ThemeManager::set_show_keybinding_hints,
-    );
-    content.append(&row);
-    append_heading(&content, "NAVIGATION");
-    for (label, keys) in [
-        ("Move through items", "↑ / ↓ (← / → in Icons)"),
-        ("Jump to top / bottom", "Ctrl + ↑ / Ctrl + ↓"),
-        ("Open item", "Enter"),
-        ("Go to parent", "Alt + ↑"),
-        ("Back / forward", "Alt + ← / →"),
-        ("Move between column panes", "← / → (Columns)"),
-        ("Focus pane navigation header", "↑ at top"),
-        ("Sidebar to top navigation bar", "↑ from sidebar top"),
-        ("Return from header to files", "↓"),
-        ("Edit location", "Ctrl + L"),
-        ("Filter items", "Ctrl + F"),
-        ("Toggle sidebar", "Ctrl + B"),
-    ] {
-        append_keybinding(&content, label, keys);
-    }
-
-    append_heading(&content, "VIEW");
-    append_keybinding(&content, "Toggle hidden files", "Ctrl + H  or  Ctrl + .");
-
-    append_heading(&content, "FILE OPERATIONS");
-    for (label, keys) in [
-        ("Create new folder", "Ctrl + Shift + N"),
-        ("Cut", "Ctrl + X"),
-        ("Copy", "Ctrl + C"),
-        ("Paste", "Ctrl + V"),
-        ("Rename", "F2 / Ctrl + R"),
-    ] {
-        append_keybinding(&content, label, keys);
-    }
-
-    append_heading(&content, "APPLICATION");
-    for (label, keys) in [
-        ("Search", "Ctrl + K"),
-        ("Open terminal", "Ctrl + T"),
-        ("Refresh", "F5"),
-        ("Open settings", "Ctrl + ,"),
-        ("Shortcut reference", "F1"),
-    ] {
-        append_keybinding(&content, label, keys);
-    }
-
-    scrollable_page(&content, Some("settings-keybindings-scroll"))
-}
-
-fn about_page() -> gtk::Widget {
-    let content = page_content();
-    content.add_css_class("about-page");
-
-    let identity = gtk::Box::new(gtk::Orientation::Vertical, 7);
-    identity.add_css_class("about-identity");
-    identity.set_halign(gtk::Align::Center);
-
-    let name = gtk::Label::new(Some("Strata"));
-    name.add_css_class("about-name");
-    let description = gtk::Label::new(Some(crate::build_info::DESCRIPTION));
-    description.add_css_class("about-description");
-    description.set_justify(gtk::Justification::Center);
-    description.set_wrap(true);
-    identity.append(&name);
-    identity.append(&description);
-    content.append(&identity);
-
-    append_heading(&content, "BUILD INFORMATION");
-    let build = gtk::Box::new(gtk::Orientation::Vertical, 0);
-    build.add_css_class("about-details");
-    let version = crate::build_info::installed_version().to_string();
-    append_about_detail(&build, "Version", &version, false);
-    let build_kind = crate::build_info::build_kind();
-    if build_kind != services::BuildKind::Stable {
-        append_about_detail(&build, "Build", build_kind.label(), false);
-    }
-    append_about_detail(&build, "Commit", crate::build_info::COMMIT, true);
-    content.append(&build);
-
-    append_heading(&content, "PROJECT");
-    let project = gtk::Box::new(gtk::Orientation::Vertical, 0);
-    project.add_css_class("about-details");
-    append_about_detail(&project, "Author", crate::build_info::AUTHOR, false);
-
-    let repository = gtk::LinkButton::builder()
-        .uri(crate::build_info::REPOSITORY)
-        .tooltip_text("Open the Strata repository")
-        .build();
-    repository.add_css_class("about-repository");
-    let repository_content = gtk::Box::new(gtk::Orientation::Horizontal, 8);
-    let repository_label = gtk::Label::new(Some("GitHub repository"));
-    repository_label.set_xalign(0.0);
-    repository_label.set_hexpand(true);
-    repository_content.append(&repository_label);
-    repository_content.append(&crate::assets::primary_icon(icons::EXTERNAL_LINK, 16));
-    repository.set_child(Some(&repository_content));
-    project.append(&repository);
-    content.append(&project);
-
-    scrollable_page(&content, None)
-}
-
-fn append_about_detail(container: &gtk::Box, label: &str, value: &str, monospace: bool) {
-    let row = gtk::Box::new(gtk::Orientation::Horizontal, 16);
-    row.add_css_class("about-detail-row");
-    let label = gtk::Label::new(Some(label));
-    label.add_css_class("about-detail-label");
-    label.set_xalign(0.0);
-    label.set_hexpand(true);
-    let value = gtk::Label::new(Some(value));
-    value.add_css_class("about-detail-value");
-    value.set_selectable(true);
-    if monospace {
-        value.add_css_class("monospace");
-    }
-    row.append(&label);
-    row.append(&value);
-    container.append(&row);
-}
-
-fn append_keybinding(content: &gtk::Box, label: &str, keys: &str) {
-    let row = gtk::Box::new(gtk::Orientation::Horizontal, 16);
-    row.add_css_class("keybinding-row");
-    let label = gtk::Label::new(Some(label));
-    label.set_xalign(0.0);
-    label.set_hexpand(true);
-    let keys = gtk::Label::new(Some(keys));
-    keys.add_css_class("keybinding-keys");
-    row.append(&label);
-    row.append(&keys);
-    content.append(&row);
 }
 
 fn navigation_button(icon: &str, label: &str) -> (gtk::Button, gtk::Label, gtk::Box) {
